+++ tests/test_interact.py
@@ -40,7 +40,7 @@ class TestInteract(XmlRpcTestCase):
                              return_value='password').start()
         self.service.db.list.return_value = ['database']
         self.service.common.login.side_effect = [17, 51]
-        self.service.object.execute_kw.side_effect = TypeError
+        self.service.object.execute_kw.side_effect = [{}, TypeError, {}]
 
         # Launch interactive
         self.infunc.side_effect = [
@@ -56,8 +56,12 @@ class TestInteract(XmlRpcTestCase):
         expected_calls = self.startup_calls + (
             ('common.login', 'database', 'usr', 'password'),
             ('object.execute_kw', 'database', 17, 'password',
-             'ir.model.access', 'check', ('res.users', 'write'), {'context': ANY}),
+             'res.users', 'context_get', ()),
+            ('object.execute_kw', 'database', 17, 'password',
+             'ir.model.access', 'check', ('res.users', 'write')),
             ('common.login', 'database', 'gaspard', 'password'),
+            ('object.execute_kw', 'database', 51, 'password',
+             'res.users', 'context_get', ()),
         )
         self.assertCalls(*expected_calls)
         self.assertEqual(getpass.call_count, 2)
@@ -108,7 +112,7 @@ class TestInteract(XmlRpcTestCase):
         mock.patch('getpass.getpass', return_value='x').start()
         self.service.db.list.return_value = ['database']
         self.service.common.login.side_effect = [17, None]
-        self.service.object.execute_kw.side_effect = [42, {}, TypeError, 42, {}]
+        self.service.object.execute_kw.side_effect = [{}, 42, {}, TypeError, 42, {}]
 
         # Launch interactive
         self.infunc.side_effect = [
@@ -119,10 +123,12 @@ class TestInteract(XmlRpcTestCase):
         odooly.main()
 
         def usr17(model, method, *params):
-            return ('object.execute_kw', 'database', 17, 'passwd', model, method, params, {'context': ANY})
+            return ('object.execute_kw', 'database', 17, 'passwd', model, method, params)
 
         expected_calls = self.startup_calls + (
             ('common.login', 'database', 'usr', 'passwd'),
+            ('object.execute_kw', 'database', 17, 'passwd',
+             'res.users', 'context_get', ()),
             usr17('ir.model', 'search',
                   [('model', 'like', 'res.company')]),
             usr17('ir.model', 'read', 42, ('model',)),